adding message if there's not enought recipes

+++ app/controllers/menus_controller.rb
@@ -14,13 +14,17 @@ class MenusController < ApplicationController
     end
 
     def generate
-        current_user.menu.menu_recipes.destroy_all
-        recipes = 7.times.map{Recipe.all.sample}
-        if current_user.menu.menu_recipes.empty?
-            recipes.each_with_index do |recipe, index|
-                menu_recipe = current_user.menu.menu_recipes.build(recipe_id: recipe.id, day_of_the_week: "#{index + 1}")
-                menu_recipe.save
-            end
+        if current_user.recipes.size > 1
+          current_user.menu.menu_recipes.destroy_all
+          recipes = 7.times.map{Recipe.all.sample}
+          if current_user.menu.menu_recipes.empty?
+              recipes.each_with_index do |recipe, index|
+                  menu_recipe = current_user.menu.menu_recipes.build(recipe_id: recipe.id, day_of_the_week: "#{index + 1}")
+                  menu_recipe.save
+              end
+          end
+        else
+          flash.alert = "Não há receitas suficientes"
         end
         redirect_to menu_path(current_user.menu)
     end
